Rename newickFormatReader.py to input_reader.py
