+++ empress/__init__.py
@@ -14,9 +14,9 @@ from typing import List
 from abc import ABC, abstractmethod
 
 from empress.xscape.CostVector import CostVector
-from empress import newickFormatReader
-from empress.newickFormatReader import ReconInput
-from empress.newickFormatReader import getInput as read_input
+from empress import input_reader
+from empress.input_reader import ReconInput
+from empress.input_reader import getInput as read_input
 from empress.xscape.reconcile import reconcile as xscape_reconcile
 from empress.xscape.plotcostsAnalytic import plot_costs_on_axis as xscape_plot_costs_on_axis
 from empress.reconcile import recongraph_tools
