+++ test/recon_vis/test_utils.py
@@ -2,7 +2,7 @@ import unittest
 import os
 import itertools
 import shutil
-from empress import newickFormatReader
+from empress import input_reader
 from empress.recon_vis import utils
 from empress.reconcile import recongraph_tools
 from empress.histogram import histogram_brute_force
@@ -234,7 +234,7 @@ class TestUtils(unittest.TestCase):
             for newick in os.listdir(tree_size_Folder):
                 if count >= self.num_examples_to_test: break
                 if newick.startswith('.'): continue
-                recon_input =  newickFormatReader.getInput(os.path.join(tree_size_Folder, newick))
+                recon_input =  input_reader.getInput(os.path.join(tree_size_Folder, newick))
                 host_tree = recon_input.host_tree
                 parasite_tree = recon_input.parasite_tree
                 for d, t, l in itertools.product(range(1, 5), repeat=3):
